feat: moved rendezvous into internal
